Update System Game

+++ Assets/ScriptableObjectData/TritsData.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Trits Data", menuName = "ScriptableObject/Trits Data")]
 public class TritsData : ScriptableObject
 {
-    public float Extraversion;
-    public float Agreeableness;
-    public float Conscientiousness;
-    public float Neuroticism;
-    public float Openness;
+    public int Extraversion;
+    public int Agreeableness;
+    public int Conscientiousness;
+    public int Neuroticism;
+    public int Openness;
 }
